api/requirements.txt
Flask
celery
redis
Flask-SQLAlchemy
pyjwt
flask_apispec
flask_restful
flask_marshmallow
Werkzeug
markupsafe==2.0.1
python-dotenv
psycopg2-binary
marshmallow-sqlalchemy
Flask-Limiter
flask-jwt-extended
flask_jwt_extended
sqlalchemy
sqlalchemy-cockroachdb
Flask-PyMongo
pymongo[srv]
pymongo